Many new examples

+++ c++/src/ITK/MaskNegatedImageFilter/MaskNegatedImageFilter.cxx
@@ -45,7 +45,7 @@ void CreateHalfMask(ImageType::Pointer image, ImageType::Pointer mask)
   // Make the left half of the mask white and the right half black
   while(!imageIterator.IsAtEnd())
   {
-    if(imageIterator.GetIndex()[0] > regionSize[0] / 2)
+    if(static_cast<unsigned int>(imageIterator.GetIndex()[0]) > regionSize[0] / 2)
         {
         imageIterator.Set(0);
         }
