+++ c++/src/ITK/MaskNeighborhoodOperatorImageFilter/MaskNeighborhoodOperatorImageFilter.cxx
@@ -65,7 +65,7 @@ void CreateHalfMask(UnsignedCharImageType::Pointer image, UnsignedCharImageType:
   // Make the left half of the mask white and the right half black
   while(!imageIterator.IsAtEnd())
   {
-    if(imageIterator.GetIndex()[0] > regionSize[0] / 2)
+    if(static_cast<unsigned int>(imageIterator.GetIndex()[0]) > regionSize[0] / 2)
         {
         imageIterator.Set(0);
         }
